AV2

+++ Atividades Ajalmar/AV2(2022.2)/pet.c
@@ -20,11 +20,28 @@ int nextCodePet() {
     return codPet;
 }
 
+int checkCodUser(char *file_path,int TestCodUser){
+
+    int codUser=0;
+    FILE *chv=NULL;
+    chv=fopen(file_path,"rb");
+
+    fseek(chv,0,SEEK_SET);
+    fread(&codUser,sizeof(int),1,chv);
+
+    for(int i=0;i<codUser;i++){
+
+        if(TestCodUser==i+1){
+            return 0;
+        }
+    }
+    return 1;
+}
 void insertPetsUI() {
-    int codUser;
-    char *Name=malloc(sizeof(char)*sizeMaxNames);
-    char *Date=malloc(sizeof(char)*9);
-    char *Type=malloc(sizeof(char)*50);
+    int codUser, option;
+    char *Name = malloc(sizeof(char) * sizeMaxNames);
+    char *Date = malloc(sizeof(char) * 9);
+    char *Type = malloc(sizeof(char) * 50);
 
     printf("\n#-----------------------------------------#");
     printf("\n|                CADASTRO                 |");
@@ -46,9 +63,26 @@ void insertPetsUI() {
     printf("\n| > CODIGO DO DONO DO PET:                |");
     printf("\n#-----------------------------------------#");
     printf("\n > ");
-    scanf("%d",&codUser);
+    scanf("%d", &codUser);
 
-   insertPets("pets.bin",Name,Date,Type,codUser);
+    if (!checkCodUser("keyUser.chv", codUser)) {
+        printf("%d",checkCodUser("keyUser.chv", codUser));
+        insertPets("pets.bin", Name, Date, Type, codUser);
+    }
+    else {
+        printf("\n#---------------------------------------------------#");
+        printf("\n|                USUARIO NAO EXISTE                 |");
+        printf("\n#---------------------------------------------------#");
+
+        printf("\n#---------------------------------------------#");
+        printf("\n|  1 - Para voltar ao Menu de pets            |");
+        printf("\n#---------------------------------------------#");
+        printf("\n > ");
+        scanf("%d", &option);
+        if (option == 1)
+            interfacePets();
+
+    }
 }
 
 void insertPets(char *file_path,char *Name,char *Date,char *Type,int codUser){
@@ -82,11 +116,14 @@ void insertPets(char *file_path,char *Name,char *Date,char *Type,int codUser){
 
 void listPets(char *file_path){
     FILE *pets;
+    FILE *persons;
     FILE *chvPet=NULL;
     int codPet=0;
     Pets listPets;
+    person listPersons;
 
     pets=fopen(file_path,"rb");
+    persons=fopen("persons.bin","rb");
     chvPet=fopen("keyPets.chv","rb");
 
     fseek(chvPet,0,SEEK_SET);
@@ -97,7 +134,11 @@ void listPets(char *file_path){
     {
         fseek(pets, sizeof(Pets) * j, SEEK_SET);
         fread(&listPets, sizeof(Pets), 1, pets);
-        if(listPets.fileExist==1) {
+
+        fseek(persons, sizeof(person) * j, SEEK_SET);
+        fread(&listPersons, sizeof(person), 1, persons);
+
+        if(listPets.fileExist==1 && listPersons.fileExist==1 ) {
             printf("\n#-----------------------------------------#");
             printf("\n| > Codigo: %.3d                           |", listPets.codPet);
             printf("\n#-----------------------------------------#");
@@ -234,35 +275,35 @@ int compareNamePet (const void *a, const void *b) {
     return strcmp (((Pets *)a)->namePet,((Pets *)b)->namePet);
 }
 
-void orderAlfPet(char *file_path){
+void orderAlfPet(char *file_path) {
 
     Pets listPets;
-    int codPet=0;
+    int codPet = 0;
 
     FILE *Pet;
     FILE *chvPet;
 
-    Pet=fopen(file_path,"rb");
-    chvPet=fopen("keyPets.chv","rb");
+    Pet = fopen(file_path, "rb");
+    chvPet = fopen("keyPets.chv", "rb");
 
-    fread(&codPet,sizeof(int),1,chvPet);
-    printf("%d",codPet);
-    int auxCod=codPet;
-    Pets allPets[codPet-1];
-    for(int i=0;i<codPet;i++) {
+    fread(&codPet, sizeof(int), 1, chvPet);
+    printf("%d", codPet);
+    int auxCod = codPet;
+    Pets allPets[codPet - 1];
+    for (int i = 0; i < codPet; i++) {
         fseek(Pet, sizeof(Pets) * i, SEEK_SET);
-        fread(&listPets,sizeof(Pets),1,Pet);
+        fread(&listPets, sizeof(Pets), 1, Pet);
 
-        strcpy(allPets[i].namePet,listPets.namePet);
-        strcpy(allPets[i].typePet,listPets.typePet);
-        strcpy(allPets[i].datePet,listPets.datePet);
-        allPets[i].codUser=listPets.codUser;
-        allPets[i].codPet=listPets.codPet;
+        strcpy(allPets[i].namePet, listPets.namePet);
+        strcpy(allPets[i].typePet, listPets.typePet);
+        strcpy(allPets[i].datePet, listPets.datePet);
+        allPets[i].codUser = listPets.codUser;
+        allPets[i].codPet = listPets.codPet;
     }
 
     qsort(allPets, codPet, sizeof(Pets), compareNamePet);
-    printf("%d",allPets[0].namePet);
-    for(int i=0;i<auxCod;i++) {
+    printf("%d", allPets[0].namePet);
+    for (int i = 0; i < auxCod; i++) {
         printf("\n#-----------------------------------------#");
         printf("\n| > Codigo: %.3d                           |", allPets[i].codPet);
         printf("\n#-----------------------------------------#");
@@ -271,8 +312,4 @@ void orderAlfPet(char *file_path){
         printf(" > Tipo do pet: %s", allPets[i].typePet);
         printf(" > Codigo do usuario: %.3d", allPets[i].codUser);
     }
-
-
-
-
 }
